feat(main-page): 為「值班教室」按鈕加上警告
為避免誤觸，按下「值班教室」按鈕後會顯示確認提示

+++ web/main.js
@@ -399,7 +399,12 @@ document.getElementById("searchBar").addEventListener("keypress", function (even
 })
 
 document.getElementById("enterGeneralClass").addEventListener("click", function (event) {
-    window.localStorage.setItem("latestClassHash", "#GENERAL");
+    dialogs.confirm("你即將進入「值班教室」模式。此模式只應該在 302 教室使用。\n是否仍要進入「值班教室」模式？", toGeneral => {
+        if (toGeneral) {
+            window.location.hash = "GENERAL";
+            window.localStorage.setItem("latestClassHash", "#GENERAL");
+        }
+    })
 })
 
 $("#clearStorageUrl").on("click", function () {
